api-service/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
django-filter==24.3
django-cors-headers==4.6.0
